requirements.txt
Flask==0.7.2
Jinja2==2.6
Werkzeug==0.7.1
simplejson==2.2.1
wsgiref==0.1.2
requests==0.10.4
facebook-sdk==0.4.0